LayoutTests/imported/w3c/web-platform-tests/speculation-rules/speculation-tags/invalid-tags.https_tag-level=ruleset&type=prefetch-expected.txt


FAIL Sec-Speculation-Tags [ruleset-based]: integer tag assert_false: expected false got true
FAIL Sec-Speculation-Tags [ruleset-based]: object tag assert_false: expected false got true
FAIL Sec-Speculation-Tags [ruleset-based]: null value tag assert_false: expected false got true
PASS Sec-Speculation-Tags [ruleset-based]: non-printable character tag
PASS Sec-Speculation-Tags [ruleset-based]: string with non-printable character tag
PASS Sec-Speculation-Tags [ruleset-based]: non-ascii string tag
PASS Sec-Speculation-Tags [ruleset-based]: 0x19 tag
FAIL Sec-Speculation-Tags [ruleset-based]: 0x20 tag assert_equals: expected "\" \"" but got ""
FAIL Sec-Speculation-Tags [ruleset-based]: 0x7E tag assert_equals: expected "\"~\"" but got "~"
PASS Sec-Speculation-Tags [ruleset-based]: 0x7F tag
